1.消息封装完善 2.具体修改 添加 BaseMessage.java					实现IMessage的基类 ByteMessage.java					内容为字节数组的message ByteUtil.java						基本类型转Byte的工具类 MessageCodecException.java			封装在message转换过程中出现的错误 MessageFactory.java					根据配置文件生成不同的message TcpByteClientHandlerTest.java		byte的测试客户端 TcpByteClientTest.java				byte的测试客户端 TestFirstByteHandler.java			byte的测试handler 修改 ConstantValue.java					将MESSAGE相关的常量添加进去 IMessage.java						在IMessage添加通用方法 INetworkConsumer.java				修改consume(StringMessage message, Channel channel)为consume(IMessage message, Channel channel) log4j-dev.xml						修改输出的日志名 MessageDecoder.java					使用接口IMessage替换原先的StringMessage MessageEncoder.java					使用接口IMessage替换原先的StringMessage NettyClientTest.java				添加一个测试方法 NetworkConsumer.java				使用接口IMessage替换原先的StringMessage server-config-dev.properties		添加一个配置项messageType ServerConfig.java					添加一个配置项messageType SessionManager.java					使用接口IMessage替换原先的StringMessage StringMessage.java TcpMessageStringHandler.java		使用接口IMessage替换原先的StringMessage TcpServerStringInitializer.java		使用接口IMessage替换原先的StringMessage TcpStringClientTest.java			使用接口IMessage替换原先的StringMessage TestFirstHandler.java				使用接口IMessage替换原先的StringMessage UserService.java					使用接口IMessage替换原先的StringMessage UserServiceImpl.java				使用接口IMessage替换原先的StringMessage,添加byte的测试方法

+++ src/main/java/com/lizhaoblog/base/concurrent/dictionary/IMessageDictionary.java
@@ -27,11 +27,11 @@ public interface IMessageDictionary {
    * @param messageId
    * @param handler
    */
-  void register(int messageId,  Class<? extends IHandler> handler);
+  void register(Short messageId,  Class<? extends IHandler> handler);
   /**
    * 根据messageId获取handler
    * @param messageId
    * @return
    */
-  IHandler getHandlerFromMessageId(Integer messageId);
+  IHandler getHandlerFromMessageId(Short messageId);
 }
